requirements.txt
aiogram==3.2.0
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
sqlite3-binary